+++ CHANGELOG.rst
@@ -1,3 +1,19 @@
+v1.4.1
+======
+Compatibility fixes
+
+Changed
+-------
+- We now use the new ``convert_coordinate`` method that has been introduced
+  with psyplot v1.4.1 (see
+  `psyplot/psyplot#39 <https://github.com/psyplot/psyplot/pull/39>`__ and
+  `#30 <https://github.com/psyplot/psy-simple/pull/30>`__)
+
+Fixed
+-----
+- psy-simple is now compatible with matplotlib 3.5 (see
+  `#31 <https://github.com/psyplot/psy-simple/pull/31>`__)
+
 v1.4.0
 ======
 Compatibility fixes and LGPL license